# specs/openapi.yaml
openapi: 3.1.0
info:
  title: YouTube Curator API
  version: 0.2.0
  description: API privada para la aplicación personal YouTube Curator.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /categories/{categoryId}/exploration-topics:
    parameters:
      - $ref: "#/components/parameters/CategoryId"
    get:
      tags: [Categories, Discovery]
      summary: Listar temas adyacentes de una categoría
      operationId: listExplorationTopics
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending, approved, rejected]
            default: all
      responses:
        "200":
          description: Temas manuales y propuestos automáticamente.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ExplorationTopic"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Categories, Discovery]
      summary: Crear y aprobar un tema adyacente manual
      operationId: createExplorationTopic
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExplorationTopicInput"
      responses:
        "201":
          description: Tema manual creado con estado `approved`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplorationTopic"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /categories/{categoryId}/exploration-topics/{topicId}:
    parameters:
      - $ref: "#/components/parameters/CategoryId"
      - $ref: "#/components/parameters/ExplorationTopicId"
    patch:
      tags: [Categories, Discovery]
      summary: Aprobar, rechazar o reactivar un tema adyacente
      operationId: updateExplorationTopic
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, approved, rejected]
                weight:
                  type: number
                  minimum: 0
                  maximum: 10
      responses:
        "200":
          description: Tema actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplorationTopic"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /discoveries:
    get:
      tags: [Discovery]
      summary: Listar el lote actual de descubrimiento
      operationId: listDiscoveries
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: band
          in: query
          schema:
            type: string
            enum: [all, related, adjacent, exploratory]
            default: all
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Candidatos activos del último lote aplicable, ordenados de forma estable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoveryPage"
        "400":
          $ref: "#/components/responses/BadRequest"
  /discoveries/{videoId}/feedback:
    post:
      tags: [Discovery]
      summary: Registrar feedback de descubrimiento
      operationId: submitDiscoveryFeedback
      parameters:
        - $ref: "#/components/parameters/VideoId"
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categoryId, action]
              properties:
                categoryId:
                  type: integer
                action:
                  type: string
                  enum:
                    - more_like_this
                    - less_like_this
                    - hide_video
                    - block_channel
                    - accept_channel
      responses:
        "200":
          description: Feedback aplicado.
          content:
            application/json:
              schema:
                type: object
                required: [applied]
                properties:
                  applied:
                    type: boolean
        "404":
          $ref: "#/components/responses/NotFound"
  /discoveries/{videoId}/hidden:
    delete:
      tags: [Discovery]
      summary: Restaurar un video oculto en una categoría
      operationId: restoreHiddenDiscovery
      parameters:
        - $ref: "#/components/parameters/VideoId"
        - name: categoryId
          in: query
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "204":
          description: Ocultación revertida. El video podrá participar en actualizaciones futuras.
        "404":
          $ref: "#/components/responses/NotFound"
  /refresh-runs:
    get:
      tags: [Refresh]
      summary: Listar actualizaciones
      operationId: listRefreshRuns
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Historial.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/RefreshRun"
    post:
      tags: [Refresh]
      summary: Solicitar actualización manual
      operationId: startRefresh
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                stages:
                  type: array
                  items:
                    type: string
                    enum:
                      - subscriptions
                      - followed_videos
                      - discovery
      responses:
        "202":
          description: Actualización iniciada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshRun"
        "409":
          $ref: "#/components/responses/Conflict"
  /refresh-runs/{refreshRunId}:
    get:
      tags: [Refresh]
      summary: Consultar actualización
      operationId: getRefreshRun
      parameters:
        - name: refreshRunId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Estado de actualización.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshRun"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema:
        type: integer
...
    VideoId:
      name: videoId
      in: path
      required: true
      schema:
        type: integer
    ExplorationTopicId:
      name: topicId
      in: path
      required: true
      schema:
        type: integer
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 30
  responses:
    BadRequest:
      description: Parámetros inválidos.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Recurso inexistente.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflicto de estado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Error de validación.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    ExplorationTopicInput:
      type: object
      required: [term]
      properties:
        term:
          type: string
          minLength: 1
          maxLength: 120
        weight:
          type: number
          minimum: 0
          maximum: 10
          default: 1
    ExplorationTopic:
      allOf:
        - $ref: "#/components/schemas/ExplorationTopicInput"
        - type: object
          required: [id, categoryId, source, status, createdAt, updatedAt]
          properties:
            id:
              type: integer
            categoryId:
              type: integer
            source:
              type: string
              enum: [manual, automatic]
            status:
              type: string
              enum: [pending, approved, rejected]
            rationale:
              type: [string, "null"]
              maxLength: 500
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    Channel:
      type: object
      required:
        - id
        - youtubeChannelId
        - title
        - subscribed
        - locallyFollowed
        - blocked
        - categoryIds
      properties:
        id:
          type: integer
        youtubeChannelId:
          type: string
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: [string, "null"]
          format: uri
        subscribed:
          type: boolean
        locallyFollowed:
          type: boolean
        blocked:
          type: boolean
        categoryIds:
          type: array
          items:
            type: integer
...
    Video:
      type: object
      required:
        - id
        - youtubeVideoId
        - channel
        - title
        - publishedAt
        - origin
        - watched
      properties:
        id:
          type: integer
        youtubeVideoId:
          type: string
        channel:
          $ref: "#/components/schemas/Channel"
        title:
          type: string
        description:
          type: string
        publishedAt:
          type: string
          format: date-time
        durationSeconds:
          type: [integer, "null"]
        thumbnailUrl:
          type: [string, "null"]
          format: uri
        contentType:
          type: string
          enum: [video, live, upcoming, unknown]
        origin:
          type: string
          enum: [followed, discovery]
        watched:
          type: boolean
        discoveryContexts:
          type: array
          items:
            $ref: "#/components/schemas/DiscoveryInfo"
...
    DiscoveryInfo:
      type: object
      required: [categoryId, band, label, score, selectionRank, reasons]
      properties:
        categoryId:
          type: integer
        band:
          type: string
          enum: [related, adjacent, exploratory]
        label:
          type: string
          enum: [Relacionado, Tema cercano, Para explorar]
        score:
          type: number
          minimum: 0
          maximum: 100
        selectionRank:
          type: integer
          minimum: 1
        reasons:
          type: array
          minItems: 1
          maxItems: 3
          items:
            type: string
    DiscoveryRecommendation:
      type: object
      required: [video, context]
      properties:
        video:
          $ref: "#/components/schemas/Video"
        context:
          $ref: "#/components/schemas/DiscoveryInfo"
    DiscoveryPage:
      type: object
      required: [items, batches]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DiscoveryRecommendation"
        batches:
          type: array
          description: Resumen del lote más reciente para cada categoría incluida en la respuesta.
          items:
            $ref: "#/components/schemas/DiscoveryBatchSummary"
        nextCursor:
          type: [string, "null"]
    DiscoveryBatchSummary:
      type: object
      required:
        - categoryId
        - refreshRunId
        - generatedAt
        - targetTotal
        - selectedTotal
        - targetByBand
        - selectedByBand
      properties:
        categoryId:
          type: integer
        refreshRunId:
          type: integer
        generatedAt:
          type: string
          format: date-time
        targetTotal:
          type: integer
          default: 8
        selectedTotal:
          type: integer
          minimum: 0
        targetByBand:
          $ref: "#/components/schemas/DiscoveryBandCounts"
        selectedByBand:
          $ref: "#/components/schemas/DiscoveryBandCounts"
        shortfallReason:
          type: [string, "null"]
          enum:
            - insufficient_candidates
            - insufficient_signals
            - no_approved_topics
            - budget_exhausted
            - quota_exhausted
            - external_error
            - null
    DiscoveryBandCounts:
      type: object
      required: [related, adjacent, exploratory]
      properties:
        related:
          type: integer
          minimum: 0
        adjacent:
          type: integer
          minimum: 0
        exploratory:
          type: integer
          minimum: 0
...
    RefreshRun:
      type: object
      required:
        - id
        - status
        - stages
        - requestedAt
        - counters
        - errors
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, running, succeeded, partial, failed]
        currentStage:
          type: [string, "null"]
        stages:
          type: array
          items:
            type: string
        requestedAt:
          type: string
          format: date-time
        startedAt:
          type: [string, "null"]
          format: date-time
        finishedAt:
          type: [string, "null"]
          format: date-time
        counters:
          type: object
          additionalProperties:
            type: object
            additionalProperties: true
        errors:
          type: array
          items:
            type: object
            required: [stage, message]
            properties:
              stage:
                type: string
              message:
                type: string
        heartbeatAt:
          type: [string, "null"]
          format: date-time
        leaseExpiresAt:
          type: [string, "null"]
          format: date-time
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true